+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Fixes
+
+- Fixed the double application of styles when a custom svg is passed into the `Icon` component.
+
 ## 4.1.4 (February 26, 2026)
 
 ### Updates